lab8 assignment strings and data types
The program will sort the argument into two variables. One of the variables will contain the digits and one of the variables will contain the alpha characters 
>>>yourfunctionname("This23IsMy7Argument5") 
It will return the entire argument to the screen, and then will sort the digits and the alpha characters in order like 2375 and ThisIsMyArgument and then report them to the screen. It will then add the digits as integers and return the sum of 2 and 3 and 7 and 5 (17)

Example Command Entered and Output Formatted Exactly as Specified:

>>>yourfunctionname("This23IsMy7Argument5")
You entered: This23IsMy7Argument5 
The Digits Were: 2375 
The Alpha Characters Were: ThisIsMyArgument 
The Sum of the Digits is: 17