project 2
Create a Python program is JES that allows the user to choose two files using the mouse and then manipulates and shows the final manipulated image(s) to the user when it is done with the manipulation. (Note: It does not take any argument at the command line but gets it's input from two files using the mouse.) You MUST supply the two images in your directory for this project.

The program manipulates the image in the following way:

It has the user choose 2 files that you supply to make your final image and your program uses both of these images to construct the final collaged image
It uses at least 3 of the following techniques found in Ch 6:
mirroring pixels
changing colors on the copied image
rotating an image
streching or changing the scale of an image
blurring an image
blending 2 images
drawing into an image
The program then displays the final image to the user.

Your final image will have no white space in it.